Run skopeo and oci-image-tool directory from the faas server
And build the function's config.json for each run of a function.
There are interim changes to faaskrun.sh that pulled the function container and unpacked
it before before running krun.  Since the faas server now pulls the contain, unpacks it,
and starts krun directly the faaskrun.sh script is now unused.